--- src/Hooks/Queries/get-likedBtn.js
@@ -1,6 +1,0 @@
-import { useMutation } from '@tanstack/react-query';
-import ProductApi from 'Apis/productApi';
-
-export const useLikedBtn = () => {
-	return useMutation(ProductApi.likedBtn);
-};
