--- src/Hooks/Queries/get-soldOut.js
@@ -1,6 +1,0 @@
-import { useMutation } from '@tanstack/react-query';
-import ProductApi from 'Apis/productApi';
-
-export const useSoldOutMutation = () => {
-	return useMutation(data => ProductApi.soldOut(data.prod_idx, data.token));
-};
